+++ L2J_Mobius_2.5_Underground/java/com/l2jmobius/gameserver/network/serverpackets/ExPetInfo.java
@@ -281,7 +281,7 @@ public class ExPetInfo extends AbstractMaskPacket<NpcInfoType>
 		if (containsMask(NpcInfoType.SPEED_MULTIPLIER))
 		{
 			packet.writeE((float) _summon.getStat().getMovementSpeedMultiplier());
-			packet.writeE(_summon.getStat().getAttackSpeedMultiplier());
+			packet.writeE((float) _summon.getStat().getAttackSpeedMultiplier());
 		}
 		if (containsMask(NpcInfoType.EQUIPPED))
 		{
